refactor(route_engine): move common classes to base
